+++ src/dialog-settings.ts
@@ -92,6 +92,14 @@ export interface DialogSettings {
    * When invoked the function is passed the dialog container and the dialog overlay elements.
    */
   position?: (dialogContainer: Element, dialogOverlay?: Element) => void;
+
+  /**
+   * This function is called when a dialog closes to restore focus to the last
+   * element that was focused when the dialog opened. It can be overridden in
+   * general settings, or on a case by case basis by providing an override when
+   * a particular dialog is opened.
+   */
+  restoreFocus?: (lastActiveElement: HTMLElement) => void;
 }
 
 /**
@@ -105,4 +113,5 @@ export class DefaultDialogSettings implements DialogSettings {
   public rejectOnCancel = false;
   public ignoreTransitions = false;
   public position?: (dialogContainer: Element, dialogOverlay: Element) => void;
+  public restoreFocus = (lastActiveElement: HTMLElement) => lastActiveElement.focus();
 }