+++ src/resources/ux-dialog.ts
@@ -1,0 +1,12 @@
+export class UxDialog {
+
+  /**
+   * @internal
+   */
+  public static $view = `<template><slot></slot></template>`;
+
+  /**
+   * @internal
+   */
+  public static $resource = 'ux-dialog';
+}